Changed views names to PascalCase
