+++ src/views/Patient/DashboardPatient.js
@@ -19,11 +19,15 @@ import ListItemText from '@mui/material/ListItemText';
 
 
 //Router imports
-import { Outlet } from 'react-router-dom';
+import { Outlet, useNavigate } from 'react-router-dom';
 
 //Icon Imports
 import AccountBoxIcon from '@mui/icons-material/AccountBox';
-import Diversity1OutlinedIcon from '@mui/icons-material/Diversity1Outlined';
+import MedicalServicesOutlinedIcon from '@mui/icons-material/MedicalServicesOutlined';
+import HistoryToggleOffOutlinedIcon from '@mui/icons-material/HistoryToggleOffOutlined';
+import FavoriteBorderOutlinedIcon from '@mui/icons-material/FavoriteBorderOutlined';
+import SaveAltOutlinedIcon from '@mui/icons-material/SaveAltOutlined';
+import QuestionAnswerOutlinedIcon from '@mui/icons-material/QuestionAnswerOutlined';
 
 const drawerWidth = 240;
 
@@ -96,8 +100,33 @@ const Drawer = styled(MuiDrawer, { shouldForwardProp: (prop) => prop !== 'open' 
 
 const servicesMenuItems = [
     {
-        text: 'Pacienți',
-        icon: <Diversity1OutlinedIcon />
+        text: 'Consultație',
+        icon: <MedicalServicesOutlinedIcon />,
+        to: '/patient/appointment'
+    },
+
+    {
+        text: 'Istoric Medical',
+        icon: <HistoryToggleOffOutlinedIcon />,
+        to: '/patient/history'
+    },
+
+    {
+        text: 'Stare curentă',
+        icon: <FavoriteBorderOutlinedIcon />,
+        to: '/patient/status'
+    },
+
+    {
+        text: 'Solicită acte',
+        icon: <SaveAltOutlinedIcon />,
+        to: '/patient/document'
+    },
+
+    {
+        text: 'Întreabă-ți doctorul',
+        icon: <QuestionAnswerOutlinedIcon />,
+        to: '/patient/chat'
     }
 ];
 
@@ -108,9 +137,10 @@ const accountMenuItems = [
     }
 ];
 
-const Dashboard_Medic = () => {
+const DashboardPatient = () => {
   const theme = useTheme();
   const [open, setOpen] = React.useState(false);
+  const navigate = useNavigate();
 
   const handleDrawerOpen = () => {
     setOpen(true);
@@ -138,7 +168,7 @@ const Dashboard_Medic = () => {
             <MenuIcon />
           </IconButton>
           <Typography variant="h6" noWrap component="div">
-            Medic Dashboard
+            Patient Dashboard
           </Typography>
         </Toolbar>
       </AppBar>
@@ -150,7 +180,7 @@ const Dashboard_Medic = () => {
         </DrawerHeader>
         <Divider />
         <List>
-          {servicesMenuItems.map(({text, icon}, index) => (
+          {servicesMenuItems.map(({text, icon, to}, index) => (
             <ListItem key={text} disablePadding sx={{ display: 'block' }}>
               <ListItemButton
                 sx={{
@@ -158,6 +188,7 @@ const Dashboard_Medic = () => {
                   justifyContent: open ? 'initial' : 'center',
                   px: 2.5,
                 }}
+                onClick={() => navigate(to)}
               >
                 <ListItemIcon
                   sx={{
@@ -207,4 +238,4 @@ const Dashboard_Medic = () => {
   );
 }
 
-export default Dashboard_Medic;
+export default DashboardPatient;
